Move graphics at 3rd-party
